Fix issue with Dockerfile, and split into build/run stages
The previous Docker image did not build due to `npm ci --production` not
instaling the dev dependencies, which are used by npm build running tsc
(as the Node types are missing without `@types/node`, which is a dev
dependency).

This also runs the application in a minimal Alpine image, with the bare
minimum necessary to run the application. The application is now also
run as a non-root user, following the principle of least privilege.

Ideally the output image would be smaller, but this is Node :)